Implemented CreateItem endpoint with JSoup routine & added ItemService test

+++ back/src/main/java/com/axbg/shelf/services/CollectionServiceImpl.java
@@ -1,11 +1,15 @@
 package com.axbg.shelf.services;
 
+import static com.axbg.shelf.config.Constants.UNSORTED_COLLECTION;
+
 import com.axbg.shelf.dao.CollectionDao;
 import com.axbg.shelf.entity.Collection;
+import com.axbg.shelf.exception.CustomException;
 import java.util.List;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -31,11 +35,15 @@ public class CollectionServiceImpl implements CollectionService {
 
     @Override
     public Collection findByName(final String name) {
-        return collectionDao.findByName(name).orElse(null);
+        return collectionDao.findByName(name, userService.findUser()).orElse(null);
     }
 
     @Override
-    public void deleteByName(String name) {
+    public void deleteByName(String name) throws CustomException {
+        if (name.equals(UNSORTED_COLLECTION)) {
+            throw new CustomException("Unsorted list cannot be deleted", HttpStatus.BAD_REQUEST);
+        }
+
         collectionDao.deleteByName(name);
     }
 }
